+++ test/test_cuda.py
@@ -1798,38 +1798,6 @@ torch.cuda.synchronize()
         loss = output.sum()
         loss.backward()
 
-    def test_autocast_custom_deprecated_warning(self):
-        class MyMM(torch.autograd.Function):
-            @staticmethod
-            @torch.cuda.amp.custom_fwd
-            def forward(ctx, x, y):
-                ctx.save_for_backward(x, y)
-                return x + y
-
-            @staticmethod
-            @torch.cuda.amp.custom_bwd
-            def backward(ctx, grad):
-                _, _ = ctx.saved_variables
-                self.assertFalse(torch.is_autocast_enabled())
-                return grad, grad
-
-        mymm = MyMM.apply
-        x = torch.randn(3, 3, requires_grad=True)
-        y = torch.randn(3, 3, requires_grad=True)
-
-        with self.assertWarnsRegex(
-            DeprecationWarning,
-            r"torch.cuda.amp.custom_fwd\(args...\) is deprecated.",
-        ):
-            with torch.amp.autocast("cuda"):
-                output = mymm(x, y)
-                loss = output.sum()
-        with self.assertWarnsRegex(
-            DeprecationWarning,
-            r"torch.cuda.amp.custom_bwd\(args...\) is deprecated.",
-        ):
-            loss.backward()
-
     def test_autocast_cat_jit(self):
         # Reported at https://github.com/pytorch/pytorch/issues/38958
