+++ torch/amp/__init__.py
@@ -2,8 +2,6 @@ from .autocast_mode import (
     _enter_autocast,
     _exit_autocast,
     autocast,
-    custom_bwd,
-    custom_fwd,
     is_autocast_available,
 )
 from .grad_scaler import GradScaler
